include/Convention.hpp - factored code from test_Analyse.cpp
