+++ test/CMakeLists.txt
@@ -38,6 +38,7 @@ set(mainProgs
 	test_Version # test project version info retrieval
 
 	test_Analyse # evaluate convention determination with simulated data
+	test_Convention # diverse conventions for representing orientations
 
 	)
